src/requirements.txt
absl-py==0.11.0
adtk==0.6.2
aiohttp==3.7.2
aioredis==1.3.1
alembic==1.4.1
altair==4.1.0
anyconfig==0.9.11
appdirs==1.4.4
appnope==0.1.0
argon2-cffi==20.1.0
arrow==0.17.0
astor==0.8.1
astunparse==1.6.3
async-timeout==3.0.1
attrs==20.3.0
azure-core==1.8.2
azure-storage-blob==12.5.0
azure-common==1.1.26
azure-servicebus==7.0.0
backcall==0.2.0
base58==2.0.1
beautifulsoup4==4.9.3
binaryornot==0.4.4
black==19.10b0
bleach==3.2.1
blessings==1.7
blinker==1.4
boto3==1.16.13
botocore==1.19.13
cachetools==4.1.1
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.3
colorful==0.5.4
cookiecutter==1.6.0
cryptography==3.2.1
cycler==0.10.0
databricks-cli==0.14.0
decorator==4.4.2
defusedxml==0.6.0
docker==4.2.2
docutils==0.15.2
entrypoints==0.3
enum-compat==0.0.3
filelock==3.0.12
flask==1.1.2
fsspec==0.6.3
future==0.18.2
gast==0.3.3
gitdb==4.0.5
gitpython==3.1.11
google==3.0.0
google-api-core==1.22.2
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
googleapis-common-protos==1.52.0
gorilla==0.3.0
gpustat==0.6.0
great-expectations==0.10.10
grpcio==1.33.2
gunicorn==20.0.4; platform_system != "Windows"
h5py==2.10.0
hiredis==1.1.0
idna==2.10
importlib-metadata==2.0.0; python_version < "3.8"
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.17.2
jeepney==0.4.3; sys_platform == "linux"
jinja2==2.11.2
jinja2-time==0.2.0
jmespath==0.10.0
joblib==0.17.0
jsonref==0.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.5
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab==0.35.6
jupyterlab-server==0.2.0
kedro==0.16.6
packages/kedro_great_expectations-0.3.0-py3-none-any.whl
keras-preprocessing==1.1.2
keyring==21.5.0
kiwisolver==1.3.1
log-symbols==0.0.14
m2r==0.2.0
mako==1.1.3
markdown==3.3.3
markupsafe==1.1.1
marshmallow==3.9.1
mistune==0.8.4
mlflow==1.11.0
msgpack==1.0.0
msrest==0.6.19
multidict==5.0.0
nbconvert==5.6.1
nbformat==5.0.8
nbstripout==0.3.9
notebook==6.1.5
numpy==1.18.5
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
opencensus==0.7.11
opencensus-context==0.1.2
opt-einsum==3.3.0
packages/optimizer-1.2.1-py3-none-any.whl
packaging==20.4
pandas==1.1.4
pandocfilters==1.4.3
parso==0.7.1
pathspec==0.8.1
pathtools==0.1.2
patsy==0.5.1
pexpect==4.8.0; python_version >= "3.3" and sys_platform != "win32" or sys_platform != "win32" or python_version >= "3.4" and sys_platform != "win32"
pickleshare==0.7.5
pillow==8.0.1
pip-tools==5.3.1
pkginfo==1.6.1
pluggy==0.13.1
poyo==0.5.0
prometheus-client==0.8.0
prometheus-flask-exporter==0.18.1
prompt-toolkit==3.0.8
protobuf==3.13.0
psutil==5.7.3
ptyprocess==0.6.0; python_version >= "3.3" and sys_platform != "win32" or sys_platform != "win32" or python_version >= "3.4" and sys_platform != "win32" or os_name != "nt" or python_version >= "3.3" and sys_platform != "win32" and (python_version >= "3.3" and sys_platform != "win32" or sys_platform != "win32" or python_version >= "3.4" and sys_platform != "win32") or python_version >= "3.4" and sys_platform != "win32" and (python_version >= "3.3" and sys_platform != "win32" or sys_platform != "win32" or python_version >= "3.4" and sys_platform != "win32")
py-spy==0.3.3
pyarrow==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydeck==0.5.0
pygments==2.7.2
pyparsing==2.4.7
pypiwin32==223; sys_platform == "win32" and python_version >= "3.6"
pyrsistent==0.17.3
python-dateutil==2.8.1
python-editor==1.0.4
python-json-logger==0.1.11
pytz==2020.4
pywin32==228; sys_platform == "win32" and python_version >= "3.6" or sys_platform == "win32" or python_version >= "3.4" and sys_platform == "win32"
pywin32-ctypes==0.2.0; sys_platform == "win32"
pywinpty==0.5.7; os_name == "nt"
pyyaml==5.3.1
pyzmq==19.0.2
qtconsole==4.7.7
qtpy==1.9.0
querystring-parser==1.2.4
ray==0.8.7
readme-renderer==28.0
redis==3.4.1
regex==2020.10.28
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
rsa==4.5; python_version != "3.4"
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2; platform_python_implementation == "CPython" and python_version < "3.9"
s3transfer==0.3.3
scikit-learn==0.23.2
scipy==1.5.4
seaborn==0.10.1
secretstorage==3.2.0; sys_platform == "linux"
semver==2.13.0
send2trash==1.5.0
shap==0.35.0
simplejson==3.17.2
six==1.12.0
smmap==3.0.4
soupsieve==2.0.1; python_version >= "3.0"
sqlalchemy==1.3.13
sqlparse==0.4.1
statsmodels==0.12.1
streamlit==0.65.2
tabulate==0.8.7
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.2
toolz==0.11.1
toposort==1.5
tornado==6.1
tqdm==4.51.0
traitlets==5.0.5
twine==3.2.0
typed-ast==1.4.1
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.25.11
validators==0.18.1
waitress==1.4.4; platform_system == "Windows"
watchdog==0.10.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
werkzeug==1.0.1
wheel==0.32.3
whichcraft==0.6.1
widgetsnbextension==3.5.1
wrapt==1.11.2
xgboost==0.90
xlrd==1.2.0
xlsxwriter==1.3.7
yarl==1.6.2
zipp==3.4.0; python_version < "3.8"
